# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Lists API",
        "title": "Lists API",
        "contact": {
            "name": "Contact ISCS-IAM",
            "email": "idev-md@groupes.epfl.ch"
        },
        "version": "1.0"
    },
    "host": "api.epfl.ch",
    "paths": {
        "/v1/trellode": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trellode"
                ],
                "summary": "Search trellode by label, unitid, type, subtype",
                "parameters": [
                    {
                        "type": "string",
                        "description": "perso",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "14290",
                        "name": "unitid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "batiments",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "batiment",
                        "name": "subtype",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "0 by default",
                        "name": "pageindex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "100 by default",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"subtype\"",
                        "description": "field to sort on (type, subtype, unit)",
                        "name": "sortcolumn",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"asc\"",
                        "description": "direction to sort",
                        "name": "sortdirection",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ListsResponse"
                        }
                    }
                }
            }
        },
        "/v1/trellode/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trellode"
                ],
                "summary": "Get list by ID or name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "105179",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.List"
                        }
                    }
                }
            }
        },
        "/v1/trellode/{id}/members": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trellode"
                ],
                "summary": "Get members of a list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "105179",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "0 by default",
                        "name": "pageindex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "100 by default",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"name\"",
                        "description": "field to sort on (name, type, subtype)",
                        "name": "sortcolumn",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"asc\"",
                        "description": "direction to sort",
                        "name": "sortdirection",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MembersResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ListsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "trellode": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.List"
                    }
                }
            }
        },
        "api.MembersResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Person"
                    }
                }
            }
        },
        "models.List": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Person"
                    }
                },
                "memberscount": {
                    "type": "integer"
                },
                "subtype": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "unitid": {
                    "type": "integer"
                },
                "usualcode": {
                    "type": "string"
                }
            }
        },
        "models.Person": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "firstnameusual": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "lastnameusual": {
                    "type": "string"
                },
                "persid": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "security": [
        {
            "BasicAuth": []
        }
    ]
}